Refactoring in samples
