+++ src/Samples/Common/ViewModels/FeatureSamples/CustomPrimitiveTypes/ITypeId.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace DotVVM.Samples.Common.ViewModels.FeatureSamples.CustomPrimitiveTypes
+{
+    public interface ITypeId
+    {
+        Guid IdValue { get; }
+    }
+
+}
+
