+++ src/Samples/Common/ViewModels/FeatureSamples/CustomPrimitiveTypes/SampleItem.cs
@@ -1,0 +1,10 @@
+﻿namespace DotVVM.Samples.Common.ViewModels.FeatureSamples.CustomPrimitiveTypes
+{
+    public class SampleItem
+    {
+        public SampleId Id { get; set; }
+        public string Text { get; set; }
+    }
+
+}
+
